requirements.txt
alembic==1.7.3
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
zipp==3.6.0
